Fix validation logic
Deleted a bug where on input of a number and then deleting it you couldn't input it again.
Fixed it by adding board[row][column] = input in the changeCellValue function which changes the value in the board array.
The problem was that even though the state was changing, the value in the board array was still the same
therefore caused a problem as in other validation functions we check whether the input value is present in the board, not the state.