Add square validation
I've added square validation function. It uses 2 helper functions to give indexes of rows and columns based on the cell's placement.
It seems to work properly. I've decided to not create another function for creating square array and instead I've put it in the isInSquare function as it's not too much code and it seems rather understandable despite having two for loops.