Add buttons
Added buttons to change the difficulty level and to check wether the sudoku is solved properly etc.
Functions to buttons are yet to be added. With the difficulty component came a new state of difficulty as well as new function to check the board for casual level and normal level.